Fix runtime memory error: compile kernel with -fno-PIC flags
When kernel runs without -fno-PIC flags, all the member of
struct is zero. The -fno-PIC flag is the opposite of -fpic,
which means:
""
-fpic

Generate position-independent code (PIC) suitable for use in a shared
library, if supported for the target machine. Such code accesses all
constant addresses through a global offset table (GOT). The dynamic
loader resolves the GOT entries when the program starts (the dynamic
loader is not part of GCC; it is part of the operating system). If
the GOT size for the linked executable exceeds a machine-specific
maximum size, you get an error message from the linker indicating
that -fpic does not work; in that case, recompile with -fPIC instead.
(These maximums are 8k on the SPARC, 28k on AArch64 and 32k on the
m68k and RS/6000. The x86 has no such limit.)

Position-independent code requires special support, and therefore
works only on certain machines. For the x86, GCC supports PIC for
System V but not for the Sun 386i. Code generated for the IBM RS/6000
is always position-independent.

When this flag is set, the macros __pic__ and __PIC__ are defined to 1.
""

Signed-off-by: Wang Xingang <wxg15527063842@gmail.com>